+++ CHANGELOG.md
@@ -1,18 +1,35 @@
 # Changelog
 
 ## Next Release
-* Added support for ONNX networks with tanh nodes to command-line interface
-* Added proof producing versions of Sign, Max, Absolute Value and Disjunction constraints
-* Added support for properties provided in VNN-LIB format for ONNX networks via the Python API
-* Fixed bugs when parsing Sigmoid layers in the C++ ONNX parser.
-* Supported additional non-linear constraints Softmax and Bilinear
-* Removed dependency on torch and drop support for Python3.7
-* Bumped ONNX version to >= 1.15.0
-* Added support for Leaky ReLU
-* Added support for Pythonic API
-* Command-line ONNX parser now supports networks with multiple outputs.
-* Command-line ONNX parser now supports the following operators: Cast, Squeeze.
-* Errors now are printed on `stderr` rather than `stdout`
+
+* Dependency changes:
+  - Dropped support for Python 3.7
+  - Now use ONNX 1.15.0 (up from 0.12.0) in both C++ and Python backends.
+  - The class `MarabouONNXNetwork` no longer depends on `torch` in Python backend.
+
+* Marabou now prints errors on `stderr` rather than `stdout`
+
+* Added proof producing versions of `Sign`, `Max`, `Absolute Value` and `Disjunction` constraints.
+
+* Changes to command-line ONNX support:
+  - Fixed bug with variable lower bounds not being set correctly.
+  - Fixed bug with sigmoid operators not being parsed correctly.
+  - Added support for `Tanh`, `Squeeze` and `Cast` operators.
+  - Added support for networks with multiple outputs
+
+* Added command-line support for properties in the VNNLIB format.
+
+* Changes to Python ONNX support:
+  - Added support for `Softmax`, `Bilinear` and `LeakyRelu` operators.
+  - `MarabouONNXNetwork` no longer exposes the fields `madeGraphEquations`, `varMap`, `constantMap`, `shapeMap`
+    as these were supposed to be internal implementation details.
+  - `MarabouONNXNetwork` no longer has a `shallowCopy` method. Instead of calling this method,
+    you should set the new parameter `preserveExistingConstraints` in the method `readONNX` to
+    `True` which has the same effect.
+
+* Added support for creating constraints using the overloaded syntax `<=`, `==` etc. in
+  the Python backend. See `maraboupy/examples/7_PythonicAPI.py` for details.
 
 ## Version 1.0.0
+
 * Initial versioned release
