+++ maraboupy/test/test_equation.py
@@ -127,4 +127,3 @@ def load_network():
     network.setUpperBound(inputVars[1], 10.0)
 
     return network
-
